Move wasm execution into call frame
