+++ radix-engine/src/model/validated_package.rs
@@ -1,11 +1,9 @@
-use sbor::rust::boxed::Box;
 use sbor::rust::collections::HashMap;
 use sbor::rust::string::String;
 use sbor::rust::vec::Vec;
 use sbor::*;
 use scrypto::abi::BlueprintAbi;
 use scrypto::buffer::scrypto_decode;
-use scrypto::core::ScryptoActorInfo;
 use scrypto::prelude::{PackageAddress, PackagePublishInput};
 use scrypto::values::ScryptoValue;
 
@@ -68,48 +66,6 @@ impl ValidatedPackage {
                 Ok(ScryptoValue::from_typed(&package_address))
             }
             _ => Err(MethodNotFound(method_name.to_string())),
-        }
-    }
-
-    pub fn invoke<'borrowed, 's, S, W, I>(
-        &self,
-        actor: &ScryptoActorInfo,
-        fn_ident: &str,
-        input: ScryptoValue,
-        system_api: &mut S,
-    ) -> Result<ScryptoValue, RuntimeError>
-    where
-        S: SystemApi<'borrowed, W, I>,
-        W: WasmEngine<I>,
-        I: WasmInstance,
-    {
-        let wasm_metering_params = system_api.fee_table().wasm_metering_params();
-        let instrumented_code = system_api
-            .wasm_instrumenter()
-            .instrument(&self.code, &wasm_metering_params);
-        let mut instance = system_api.wasm_engine().instantiate(&instrumented_code);
-        let blueprint_abi = self
-            .blueprint_abi(actor.blueprint_name())
-            .expect("Blueprint should exist");
-        let export_name = &blueprint_abi.get_fn_abi(fn_ident).unwrap().export_name;
-        let mut runtime: Box<dyn WasmRuntime> =
-            Box::new(RadixEngineWasmRuntime::new(actor.clone(), system_api));
-        let output = instance
-            .invoke_export(export_name, &input, &mut runtime)
-            .map_err(|e| match e {
-                // Flatten error code for more readable transaction receipt
-                InvokeError::RuntimeError(e) => e,
-                e @ _ => RuntimeError::InvokeError(e.into()),
-            })?;
-
-        let fn_abi = blueprint_abi.get_fn_abi(fn_ident).unwrap();
-        if !fn_abi.output.matches(&output.dom) {
-            Err(RuntimeError::InvalidFnOutput {
-                fn_ident: fn_ident.to_string(),
-                output: output.dom,
-            })
-        } else {
-            Ok(output)
         }
     }
 }
