+++ radix-engine/src/wasm/wasm_instrumenter.rs
@@ -20,7 +20,7 @@ impl WasmInstrumenter {
         &mut self,
         code: &[u8],
         wasm_metering_params: &WasmMeteringParams,
-    ) -> Vec<u8> {
+    ) -> &[u8] {
         let code_hash = hash(code);
         self.cache
             .entry((code_hash, wasm_metering_params.identifier()))
@@ -34,6 +34,5 @@ impl WasmInstrumenter {
                     .expect("Failed to instrument code")
                     .0
             })
-            .clone()
     }
 }
